+++ src/components/Common/Aboutus.tsx
@@ -3,12 +3,7 @@ import { Link } from 'react-router-dom';
 import LaxButton from '../Shared/LaxButton';
 import lax from 'lax.js';
 
-import signature from "../../assets/images/signature.png"
-import about1 from "../../assets/images/about1.jpg"
-import about2 from "../../assets/images/about2.jpg"
-import about3 from "../../assets/images/about3.jpg"
-
-const About = () => {
+const About = (props: any) => {
   useEffect(() => {
     lax.setup();
 
@@ -23,7 +18,7 @@ const About = () => {
       document.removeEventListener('scroll', handleScroll);
     };
   }, []);
-  
+
   return (
     <section className="about-area ptb-120 bg-image">
       <div className="container">
@@ -32,13 +27,10 @@ const About = () => {
             <div className="about-content">
               <span>Join The Event</span>
               <h2>We Create and <b>Turn</b> Into Reality</h2>
-
-              <p>It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged.</p>
-
+              <p>Ithas survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged.</p>
               <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.</p>
-
               <div className="signature">
-                <img src={signature} alt="signature" />
+                <img src={props.signature} alt="signature" />
               </div>
 
               <Link to="/about" className="btn btn-primary">
@@ -52,12 +44,9 @@ const About = () => {
 
           <div className="col-lg-6">
             <div className="about-image">
-              <img src={about1} className="about-img1" alt="about" />
-
-              <img src={about2} className="about-img2" alt="about" />
-
-              <img src={about3} className="shape-img" alt="about" />
-
+              <img src={props.about1} className="about-img1" alt="about" />
+              <img src={props.about2} className="about-img2" alt="about" />
+              <img src={props.about3} className="shape-img" alt="about" />
               <LaxButton buttonText="Explore More About" />
             </div>
           </div>
