+++ src/components/Common/PagesMainBanner.tsx
@@ -1,21 +1,21 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-function MainBanner() {
+function PagesMainBanner(props: any) {
   return (
-    <div className="page-title-area item-bg3">
+    <div className="page-title-area item-bg1">
       <div className="container">
-        <h1>Pricing</h1>
-        <span>Get Your Tickets</span>
+        <h1>{props.title}</h1>
+        <span>{props.subTitle}</span>
         <ul>
           <li>
             <Link to="/">Home</Link>
           </li>
-          <li>Pricing</li>
+          <li>{props.path}</li>
         </ul>
       </div>
     </div>
   );
 }
 
-export default MainBanner;
+export default PagesMainBanner;
