+++ src/components/pages/AboutOne.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
-import MainBanner from "../AboutOne/MainBanner";
-import About from "../AboutOne/About";
+import PagesMainBanner from "../Common/PagesMainBanner";
+import About from "../Common/Aboutus";
 import WhyUs from "../Common/WhyUs";
 import Speakers from "../AboutOne/Speakers";
 import GoTop from "../Shared/GoTop";
@@ -11,6 +11,11 @@ import BuyTicket from "../Common/BuyTicket";
 import Subscribe from "../Common/Subscribe";
 import Cta from "../Common/Cta";
 import lax from 'lax.js';
+
+import signature from "../../assets/images/signature.png"
+import about1 from "../../assets/images/about1.jpg"
+import about2 from "../../assets/images/about2.jpg"
+import about3 from "../../assets/images/about3.jpg"
 
 function AboutOne() {
   useEffect(() => {
@@ -31,9 +36,9 @@ function AboutOne() {
   return (
     <>
       {/* Main Banner */}
-      <MainBanner />
+      <PagesMainBanner title="About Us" subTitle="Are you ready to attend?" path="About" />
       {/* About Area */}
-      <About />
+      <About about1={about1} about2={about2} about3={about3} signature={signature} />
       {/* Why Choose Us Area */}
       <WhyUs />
       {/* Speakers Area */}
