--- src/components/pages/SpeakersOne.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import MainBanner from "../Speakers/MainBanner";
-import Speakers from "../Speakers/SpeakersOne";
-import Footer from "../Common/Footer";
-
-function SpeakersOne() {
-  return (
-    <>
-      <MainBanner />
-      <Speakers />
-      <Footer />
-    </>
-  );
-}
-
-export default SpeakersOne;
